news_atlas/requirements.txt
dash
numpy
pandas
google-cloud-firestore
gunicorn